+++ hikka/inline/gallery.py
@@ -263,7 +263,7 @@ class Gallery(InlineUnit):
                     exc = "\n".join(exc.splitlines()[1:])
                     msg = (
                         "<b>🚫 Gallery invoke failed!</b>\n\n"
-                        f"<b>🧾 Logs:</b>\n<code>{exc}</code>"
+                        f"<b>🧾 Logs:</b>\n<code>{utils.escape_html(exc)}</code>"
                     )
 
                 del self._units[unit_id]